Move library updating code from component into model. Clean up and comment custom unit code. Move mathjs-utils.ts into utils/.
