+++ src/diagram/components/ui/expandable-input.scss
@@ -36,7 +36,6 @@
     textarea {
       height: 46px;
       padding: 3px 20px 3px 21px;
-      // text-indent: -15px;
       white-space: normal;
     }
   }
