+++ src/diagram/models/variable.ts
@@ -1,9 +1,11 @@
 import math, { parse, SymbolNode } from "mathjs";
-import { IAnyComplexType, Instance, types } from "mobx-state-tree";
+import { reaction } from "mobx";
+import { addDisposer, IAnyComplexType, Instance, types } from "mobx-state-tree";
 import { nanoid } from "nanoid";
 
-import { getMathUnit, getUsedInputs, replaceInputNames } from "./mathjs-utils";
+import { getMathUnit, getUsedInputs, replaceInputNames } from "../utils/mathjs-utils";
 import { createMath } from "../custom-mathjs";
+import { customUnitsArray } from "../custom-mathjs-units";
 import { basicErrorMessage, ErrorMessage, getErrorMessage } from "../utils/error";
 import { Colors, legacyColors } from "../utils/theme-utils";
 
@@ -59,6 +61,15 @@ export const Variable = types.model("Variable", {
 .actions(self => ({
   recreateMath() {
     self.math = createMath();
+  }
+}))
+.actions(self => ({
+  afterCreate() {
+    // Update the math library whenever the custom units have changed.
+    addDisposer(self, reaction(
+      () => customUnitsArray.length,
+      () => self.recreateMath()
+    ));
   }
 }))
 .views(self => ({
