+++ src/diagram/utils/mathjs-utils.test.ts
